Added confirm_credetionals() method to UserManager which returns True if the logged-in user successfully entered their password.
Added three new password-confirmation prompts for administrators trying to add,
edit, or remove a user profile.

Addressed an edge case where an administrator deletes their own account. Since
admins are routed to /dashboard after deleting an account an error is thrown when
the admin deletes their own account (indexerror because request.session["user_id] is
still set to the admin's user id but there is no account for the admin anymore).
Handled this by revising the initialize_session() method in login_and_registration_app
to check for cases where the request.session["user_id] is not in the database.
On account deletion, admin is routed to /dashboard unless they delete their own account
in which case they are routed to index.html in the login_and_registration_app.